requirements.txt
netifaces
fastapi
uvicorn
pydantic
websockets
pytest
pytest-asyncio
redis
paho-mqtt
psutil
